Move definition of LANGUAGES to redux file
