+++ frontend/src/components/eMIB/TeamInformation.jsx
@@ -7,7 +7,7 @@ import markdown_section1_fr from "./sample_test_markdown/TeamInformation_section
 import markdown_section2_en from "./sample_test_markdown/TeamInformation_section2_en.md";
 import markdown_section2_fr from "./sample_test_markdown/TeamInformation_section2_fr.md";
 import LOCALIZE from "../../text_resources";
-import { LANGUAGES } from "../commons/Translation";
+import { LANGUAGES } from "../../modules/LocalizeRedux";
 import PopupBox, { BUTTON_TYPE } from "../commons/PopupBox";
 import emib_sample_test_example_team_chart_en from "../../images/emib_sample_test_example_team_chart_en.png";
 import emib_sample_test_example_team_chart_en_zoomed from "../../images/emib_sample_test_example_team_chart_en.png";
