+++ frontend/src/tests/Status.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import Status from "../Status";
-import { LANGUAGES } from "../components/commons/Translation";
+import { LANGUAGES } from "../modules/LocalizeRedux";
 import LOCALIZE from "../text_resources";
 
 it("renders Status Page Welcome Message in English", () => {
